refactor(closure-reopen-service): Split closure and reopen-request routes
Both features move route/middleware declarations and response Zod
schemas into their own *.routes.ts files; controllers keep only
request-handling logic via createXController(service) factories.
module.ts files now create the shared DocumentedRouter and register
routes into it.

No behavior change — verified route-for-route identical OpenAPI doc
before/after (4 routes), lint clean, all 42 existing tests pass, and
live smoke checks against the running service.
