+++ apps/closure-reopen-service/src/closures/closure.module.ts
@@ -1,14 +1,14 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { ClosureRepository } from './closure.repository';
 import { ClosureService } from './closure.service';
-import { createClosureRouter } from './closure.controller';
+import { registerClosureRoutes } from './closure.routes';
 import { ApprovalClient } from '../reopen-requests/approval.client';
 import { LookupClient } from '../reopen-requests/lookup.client';
 import { NotificationClient } from '../reopen-requests/notification.client';
 
 /**
- * Composition root for the closures feature: wires repository → service → router.
+ * Composition root for the closures feature: wires repository → service → routes.
  * Replaces the former NestJS module + DI container.
  */
 export function createClosureModule(prisma: PrismaService): DocumentedRouter {
@@ -17,5 +17,7 @@ export function createClosureModule(prisma: PrismaService): DocumentedRouter {
   const lookupClient = new LookupClient();
   const notificationClient = new NotificationClient();
   const service = new ClosureService(repository, approvalClient, lookupClient, notificationClient);
-  return createClosureRouter(service);
+  const doc = createDocumentedRouter();
+  registerClosureRoutes(doc, service);
+  return doc;
 }
