+++ apps/closure-reopen-service/src/reopen-requests/reopen-request.module.ts
@@ -1,8 +1,8 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { ReopenRequestRepository } from './reopen-request.repository';
 import { ReopenRequestService } from './reopen-request.service';
-import { createReopenRequestRouter } from './reopen-request.controller';
+import { registerReopenRequestRoutes } from './reopen-request.routes';
 import { AuditClient } from './audit.client';
 import { NotificationClient } from './notification.client';
 import { ApprovalClient } from './approval.client';
@@ -10,7 +10,7 @@ import { LookupClient } from './lookup.client';
 import { BeneficiaryClient } from './beneficiary.client';
 
 /**
- * Composition root for the reopen-requests feature: wires repository → service → router.
+ * Composition root for the reopen-requests feature: wires repository → service → routes.
  */
 export function createReopenRequestModule(prisma: PrismaService): DocumentedRouter {
   const repository = new ReopenRequestRepository(prisma);
@@ -27,5 +27,7 @@ export function createReopenRequestModule(prisma: PrismaService): DocumentedRout
     lookupClient,
     beneficiaryClient,
   );
-  return createReopenRequestRouter(service);
+  const doc = createDocumentedRouter();
+  registerReopenRequestRoutes(doc, service);
+  return doc;
 }
